add typescript datadog script and note about support (#45)
We were made aware of another example bit of this code, written in
Typescript by a customer. Add that to the repo and clarify that we
explicitly don't support the use of these apps and they are meant to be
reference material for creating best-fit applications for a users
environment.